Reduce calls to inet_aton; rewrite mask converter
Move to a single call to inet_aton.  The logic above now merely figures
out what to pass to it via base_address.  Makes the ctor look a good bit
cleaner.

Change the mask converter to return a stringified int instead of a
dotted quad.  The conversion to dotted quad will happen down the line
anyway; no need to convert multiple times.  Also put in exceptions for
non-numeric masks and masks out of the valid range.